verifier: final WiP
