+++ src/redshift/IOP/FRI/coset_combining_fri/mod.rs
@@ -8,7 +8,7 @@ use crate::SynthesisError;
 use crate::multicore::Worker;
 use crate::ff::PrimeField;
 
-use crate::redshift::IOP::oracle::{Oracle, BatchedOracle};
+use crate::redshift::IOP::oracle::{Oracle, BatchedOracle, Label};
 use crate::redshift::polynomials::*;
 use crate::redshift::IOP::channel::Channel;
 
@@ -43,7 +43,7 @@ impl<F: PrimeField, I: Oracle<F>> FriProofPrototype<F, I> {
 #[derive(PartialEq, Eq, Clone)]
 pub struct FriProof<F: PrimeField, I: Oracle<F>> {
     //for batched version of FRI
-    pub upper_layer_queries: Vec<Vec<I::Query>>,
+    pub upper_layer_queries: Vec<Vec<(Label, I::Query)>>,
     pub queries: Vec<Vec<I::Query>>,
     //excluding the initial one oracle - it is simly the commitment, no need to repeat
     pub commitments: Vec<I::Commitment>,
