+++ src/redshift/IOP/oracle/mod.rs
@@ -35,6 +35,8 @@ pub trait IopQuery<F: PrimeField>: 'static + PartialEq + Eq + Clone + std::fmt::
     fn card(&self) -> usize;
 }
 
+pub type Label = &'static str;
+
 // this struct is used as an aggregator of multiple different oracles.
 // as a common example comsider the following situation: we use batch version of FRI
 // at a "highest" level, we have series of different oracles, and we need to open them simultaneously
@@ -43,40 +45,46 @@ pub trait IopQuery<F: PrimeField>: 'static + PartialEq + Eq + Clone + std::fmt::
 pub struct BatchedOracle<'a, F, I>
 where F: PrimeField, I: Oracle<F>
 {
-    pub oracles: Vec<&'a I>,
+    pub labeled_oracles: Vec<(Label, &'a I)>,
     _marker_f: std::marker::PhantomData<F>,
 }
 
 impl<'a, F: PrimeField, I: Oracle<F>> BatchedOracle<'a, F, I>
 {
-    pub fn create(oracles: Vec<&'a I>) -> Self {
+    pub fn create(labeled_oracles: Vec<(Label, &'a I)>) -> Self {
         // all of the suboracles should have the same size
-        assert!(oracles.windows(2).all(|w| w[0].size() == w[1].size()));
+        assert!(labeled_oracles.windows(2).all(|w| w[0].1.size() == w[1].1.size()));
         // and there should be at least one oracle!
-        assert!(!oracles.is_empty());
+        assert!(!labeled_oracles.is_empty());
 
-        BatchedOracle{ oracles, _marker_f : std::marker::PhantomData::<F> }
+        BatchedOracle{ labeled_oracles, _marker_f : std::marker::PhantomData::<F> }
     }
 
     pub fn size(&self) -> usize {
-        self.oracles[0].size()
+        self.labeled_oracles[0].1.size()
     }
 
     pub fn get_commitment(&self) -> Vec<I::Commitment> {
-        self.oracles.iter().map(|x| x.get_commitment()).collect()
+        self.labeled_oracles.iter().map(|x| x.1.get_commitment()).collect()
     }
 
-    pub fn verify_query(commitment: &Vec<I::Commitment>, query: &Vec<I::Query>, params: &I::Params) -> bool {
-        for (sub_query, sub_commitment) in query.iter().zip(commitment.iter()) {
-            if !I::verify_query(sub_commitment, sub_query, params) {
+    pub fn verify_query(commitment: &Vec<(Label, I::Commitment)>, query: &Vec<(Label, I::Query)>, params: &I::Params) -> bool {
+        for (label, sub_query) in query.iter() {
+            let sub_commitment = commitment.iter().find(|(l, c)| l == label);
+            let result = match sub_commitment {
+                None => false,
+                Some((l,c)) => I::verify_query(c, sub_query, params),
+            };
+
+            if !result {
                 return false;
             }
         }
         true
     }
 
-    pub fn produce_query(&self, indexes: Range<usize>, values: &Vec<&[F]>) -> Vec<I::Query> {
-        self.oracles.iter().zip(values.iter()).map(|(x, val)| x.produce_query(indexes.clone(), val)).collect()
+    pub fn produce_query(&self, indexes: Range<usize>, values: &Vec<&[F]>) -> Vec<(Label, I::Query)> {
+        self.labeled_oracles.iter().zip(values.iter()).map(|(x, val)| (x.0, x.1.produce_query(indexes.clone(), val))).collect()
     }
 }
